Выделит логику задачи в отдельный презентер
